requirements_complete.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astor==0.8.1
async-timeout==4.0.2
attrs==23.1.0
autocfg==0.0.8
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
datasets==2.11.0
Deprecated==1.2.13
dill==0.3.6
expecttest==0.1.4
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
future==0.18.3
gluoncv==0.10.5.post0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
grpcio==1.53.0
httplib2==0.15.0
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
json-tricks==3.17.0
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
neural-compressor==2.1
nni==2.10.1
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
onnx==1.13.1
onnx-opcounter==0.0.3
onnxruntime==1.12.1
opencv-python==4.7.0.72
openmim==0.3.7
optimum==1.8.2
optimum-intel==1.7.0
ordered-set==4.1.0
packaging==23.0
pandas==2.0.0
pascal-voc-writer==0.1.4
Pillow==9.5.0
portalocker==2.7.0
prettytable==3.7.0
protobuf==3.20.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
PyDrive==1.3.1
Pygments==2.14.0
pyparsing==3.0.9
pyre-extensions==0.0.30
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PythonWebHDFS==0.2.3
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.3
rsa==4.9
schema==0.7.5
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.98
simplejson==3.19.1
six==1.16.0
sklearn==0.0.post5
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.0
torcheval==0.0.6
torchtnt==0.1.0
torchvision==0.14.0
tqdm==4.65.0
transformers==4.28.1
typeguard==2.13.3
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.15.0
xxhash==3.2.0
yacs==0.1.8
yapf==0.32.0
yarl==1.9.1
